requirements.txt
fastapi
uvicorn
pillow
opencv-python
numpy
pydantic
jinja2
